Project split into two separate modules for library and demo
