+++ android-image-zoom-crop/.gitignore
@@ -1,0 +1,1 @@
+/build
