+++ android-image-zoom-crop/build.gradle
@@ -1,11 +1,10 @@
-apply plugin: 'com.android.application'
+apply plugin: 'com.android.library'
 
 android {
     compileSdkVersion 23
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "io.togoto.imagezoomcrop.demo"
         minSdkVersion 14
         targetSdkVersion 23
         versionCode 1
@@ -21,8 +20,6 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile (project(':android-image-zoom-crop'))
+    testCompile 'junit:junit:4.12'
     compile 'com.android.support:appcompat-v7:23.1.0'
-    compile 'com.android.support:cardview-v7:23.1.0'
-    compile 'com.joooonho:selectableroundedimageview:1.0.1'
 }
