+++ android-image-zoom-crop/src/androidTest/java/io/togoto/imagezoomcrop/ApplicationTest.java
