+++ android-image-zoom-crop/src/main/AndroidManifest.xml
@@ -1,0 +1,12 @@
+<manifest
+    package="io.togoto.imagezoomcrop"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <application
+        android:allowBackup="true"
+        android:label="@string/app_name"
+        android:supportsRtl="true"
+        >
+    </application>
+
+</manifest>
