+++ android-image-zoom-crop/src/main/java/io/togoto/imagezoomcrop/imagecrop/cropoverlay/edge/EdgePair.java
