+++ android-image-zoom-crop/src/main/java/io/togoto/imagezoomcrop/imagecrop/cropoverlay/utils/PaintUtil.java
