+++ android-image-zoom-crop/src/main/java/io/togoto/imagezoomcrop/imagecrop/photoview/gestures/CupcakeGestureDetector.java
