+++ android-image-zoom-crop/src/main/res/values/attrs.xml
