+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="io.togoto.imagezoomcrop" >
+    package="io.togoto.imagezoomcrop.demo" >
 
     <!-- Allow Disk Caching on SD Card -->
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
