+++ app/src/main/java/io/togoto/imagezoomcrop/demo/GOTOConstants.java
@@ -1,4 +1,4 @@
-package io.togoto.imagezoomcrop;
+package io.togoto.imagezoomcrop.demo;
 
 /**
  * @author GT
