+++ app/src/main/java/io/togoto/imagezoomcrop/demo/ImageCropActivity.java
@@ -1,4 +1,4 @@
-package io.togoto.imagezoomcrop;
+package io.togoto.imagezoomcrop.demo;
 
 import android.app.Activity;
 import android.content.ActivityNotFoundException;
